SandboxSubmissionService defined in terms of ContractStore and WriteService
Towards #1273.
