+++ ledger/participant-state-index/src/main/scala/com/daml/ledger/participant/state/index/v2/ContractStore.scala
@@ -4,6 +4,8 @@
 package com.daml.ledger.participant.state.index.v2
 
 import com.digitalasset.daml.lf.data.Ref
+import com.digitalasset.daml.lf.data.Ref.Party
+import com.digitalasset.daml.lf.transaction.Node.GlobalKey
 import com.digitalasset.daml.lf.value.Value
 import com.digitalasset.daml.lf.value.Value.{AbsoluteContractId, ContractInst}
 
@@ -17,5 +19,7 @@ trait ContractStore {
       submitter: Ref.Party,
       contractId: AbsoluteContractId
   ): Future[Option[ContractInst[Value.VersionedValue[AbsoluteContractId]]]]
+
+  def lookupContractKey(submitter: Party, key: GlobalKey): Future[Option[AbsoluteContractId]]
 
 }
