+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/services/SandboxSubmissionService.scala
@@ -4,6 +4,7 @@
 package com.digitalasset.platform.sandbox.services
 import com.digitalasset.ledger.api.v1.command_submission_service.CommandSubmissionServiceLogging
 import akka.stream.ActorMaterializer
+import com.daml.ledger.participant.state.index.v2.ContractStore
 import com.daml.ledger.participant.state.v1.SubmissionResult.{Acknowledged, Overloaded}
 import com.daml.ledger.participant.state.v1.{
   SubmissionResult,
@@ -20,7 +21,7 @@ import com.digitalasset.daml.lf.transaction.Transaction.Transaction
 import com.digitalasset.grpc.adapter.utils.DirectExecutionContext
 import com.digitalasset.ledger.api.domain.{LedgerId, Commands => ApiCommands}
 import com.digitalasset.ledger.api.messages.command.submission.SubmitRequest
-import com.digitalasset.ledger.backend.api.v1.{LedgerBackend, TransactionSubmission}
+import com.digitalasset.ledger.backend.api.v1.TransactionSubmission
 import com.digitalasset.platform.sandbox.config.DamlPackageContainer
 import com.digitalasset.platform.sandbox.stores.ledger.{CommandExecutor, ErrorCause}
 import com.digitalasset.platform.server.api.services.domain.CommandSubmissionService
@@ -36,8 +37,6 @@ import scala.compat.java8.FutureConverters
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
 
-import com.digitalasset.ledger.api.domain.LedgerId
-
 object SandboxSubmissionService {
 
   type RecordUpdate = Either[LfError, (Transaction, BlindingInfo)]
@@ -46,7 +45,7 @@ object SandboxSubmissionService {
       ledgerId: LedgerId,
       packageContainer: DamlPackageContainer,
       identifierResolver: IdentifierResolver,
-      ledgerBackend: LedgerBackend,
+      contractStore: ContractStore,
       writeService: WriteService,
       timeModel: TimeModel,
       timeProvider: TimeProvider,
@@ -55,7 +54,7 @@ object SandboxSubmissionService {
     new GrpcCommandSubmissionService(
       new SandboxSubmissionService(
         packageContainer,
-        ledgerBackend,
+        contractStore,
         writeService,
         timeModel,
         timeProvider,
@@ -72,7 +71,7 @@ object SandboxSubmissionService {
 
 class SandboxSubmissionService private (
     packageContainer: DamlPackageContainer,
-    ledgerBackend: LedgerBackend,
+    contractStore: ContractStore,
     writeService: WriteService,
     timeModel: TimeModel,
     timeProvider: TimeProvider,
@@ -126,13 +125,12 @@ class SandboxSubmissionService private (
 
   private def recordOnLedger(commands: ApiCommands): Future[SubmissionResult] =
     for {
-      handle <- ledgerBackend.beginSubmission
       res <- commandExecutor
         .execute(
           commands.submitter,
           commands,
-          handle.lookupActiveContract(commands.submitter, _),
-          handle.lookupContractKey(commands.submitter, _),
+          contractStore.lookupActiveContract(commands.submitter, _),
+          contractStore.lookupContractKey(commands.submitter, _),
           commands.commands
         )
       submissionResult <- handleResult(res)
