+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/stores/ledger/SandboxLedgerBackend.scala
@@ -8,18 +8,11 @@ import java.util.concurrent.CompletionStage
 import akka.NotUsed
 import akka.stream.Materializer
 import akka.stream.scaladsl.Source
-import com.daml.ledger.participant.state.index.v2.{
-  AcsUpdateEvent,
-  ActiveContractSetSnapshot,
-  ActiveContractsService,
-  TransactionsService
-}
+import com.daml.ledger.participant.state.index.v2._
 import com.daml.ledger.participant.state.{v1 => ParticipantState}
 import com.digitalasset.daml.lf.data.Ref
-import com.digitalasset.daml.lf.data.Ref.{LedgerString, Party, TransactionIdString}
+import com.digitalasset.daml.lf.data.Ref.{LedgerString, TransactionIdString}
 import com.digitalasset.daml.lf.engine.Blinding
-import com.digitalasset.daml.lf.transaction.Node
-import com.digitalasset.daml.lf.transaction.Transaction.{Value => TxValue}
 import com.digitalasset.daml.lf.value.Value
 import com.digitalasset.ledger.api.domain
 import com.digitalasset.ledger.api.domain.{EventId, LedgerOffset, TransactionFilter, TransactionId}
@@ -34,58 +27,16 @@ import com.digitalasset.platform.participant.util.EventFilter
 import com.digitalasset.platform.sandbox.stores.ActiveContracts
 import com.digitalasset.platform.server.api.validation.ErrorFactories
 import com.digitalasset.platform.server.services.transaction.TransactionConversion
+import scalaz.syntax.tag._
 
 import scala.compat.java8.FutureConverters
 import scala.concurrent.{ExecutionContext, Future}
-import scalaz.syntax.tag._
 
 class SandboxLedgerBackend(ledger: Ledger)(implicit mat: Materializer)
     extends LedgerBackend //TODO: remove this later so we can rely on sole participant state interfaces
     with ParticipantState.WriteService
     with ActiveContractsService
     with TransactionsService {
-
-  private class SandboxSubmissionHandle extends SubmissionHandle {
-    override def abort: Future[Unit] = Future.successful(())
-
-    override def submit(
-        submitted: TransactionSubmission): Future[ParticipantState.SubmissionResult] =
-      ledger.publishTransaction(submitted)
-
-    private[this] def canSeeContract(
-        submitter: Party,
-        ac: ActiveContracts.ActiveContract): Boolean = {
-      // ^ only parties disclosed or divulged to can lookup; see https://github.com/digital-asset/daml/issues/10
-      // and https://github.com/digital-asset/daml/issues/751 .
-      Party fromString submitter exists (p => ac.witnesses(p) || ac.divulgences.contains(p))
-    }
-
-    override def lookupActiveContract(submitter: Party, contractId: Value.AbsoluteContractId)
-      : Future[Option[Value.ContractInst[TxValue[Value.AbsoluteContractId]]]] =
-      ledger
-        .lookupContract(contractId)
-        .map(_.collect {
-          case ac if canSeeContract(submitter, ac) => ac.contract
-        })(DEC)
-
-    override def lookupContractKey(
-        submitter: Party,
-        key: Node.GlobalKey): Future[Option[Value.AbsoluteContractId]] = {
-      implicit val ec: ExecutionContext = DEC
-      ledger.lookupKey(key).flatMap {
-        // note that we need to check visibility for keys, too, otherwise we leak the existence of a non-divulged
-        // contract if we return `Some`.
-        case None => Future.successful(None)
-        case Some(cid) =>
-          ledger.lookupContract(cid) map {
-            _ flatMap (ac => if (canSeeContract(submitter, ac)) Some(cid) else None)
-          }
-      }
-    }
-  }
-
-  override def beginSubmission(): Future[SubmissionHandle] =
-    Future.successful(new SandboxSubmissionHandle)
 
   override def ledgerSyncEvents(
       offset: Option[LedgerSyncOffset]): Source[LedgerSyncEvent, NotUsed] =
@@ -213,10 +164,7 @@ class SandboxLedgerBackend(ledger: Ledger)(implicit mat: Materializer)
         transaction
       )
 
-    val resultF = for {
-      handle <- beginSubmission()
-      result <- handle.submit(transactionSubmission)
-    } yield result
+    val resultF = ledger.publishTransaction(transactionSubmission)
 
     FutureConverters.toJava(resultF)
   }
